P157

+++ src/main/java/com/traveloveapi/repository/voucher/VoucherRepository.java
@@ -97,7 +97,7 @@ public class VoucherRepository {
         for (Object ele: data)
         {
             Object[] row = (Object[]) ele;
-            VoucherDTO a = new VoucherDTO((String) row[0], (VoucherDiscountType) row[1], (float) row[2], (double) row[3], (float) row[4], (float) row[5], (String) row[6], (VoucherTargetType) row[7], (String) row[8], currency);
+            VoucherDTO a = new VoucherDTO((String) row[0], (VoucherDiscountType) row[1], (float) row[2], ((Float)row[3]).doubleValue(), (float) row[4], (float) row[5], (String) row[6], (VoucherTargetType) row[7], (String) row[8], currency);
             if (currency!=null && a.getType()==VoucherDiscountType.FIXED)
                 a.setFixed_discount(currencyService.convert(a.getOriginCurrency(), currency, a.getFixed_discount()));
             rs.add(a);
